app/requirements.txt
Django==5.2.3
paho-mqtt==1.6.1
pytz==2023.3
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
ultralytics
sahi